방문자 리스트 뷰모델 추가 및 추출 기능 옮기기
