+++ app/src/main/AndroidManifest.xml
@@ -21,6 +21,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".visitorList.VisitorListActivity"/>
     </application>
 
 </manifest>
